requirements.txt
# Core deep learning frameworks
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0

# Data processing and scientific computing
numpy>=1.21.0
scipy>=1.9.0
scikit-learn>=1.1.0
opencv-python>=4.6.0
pillow>=9.0.0

# 3D data processing
mrcfile>=1.4.0
h5py>=3.7.0
nibabel>=4.0.0

# Visualization
matplotlib>=3.5.0
plotly>=5.10.0
seaborn>=0.11.0
vtk>=9.2.0
mayavi>=4.8.0

# Experiment tracking and management
wandb>=0.15.0
tensorboard>=2.10.0
mlflow>=2.0.0

# Configuration management
pyyaml>=6.0
hydra-core>=1.2.0
omegaconf>=2.2.0

# Data augmentation
albumentations>=1.2.0
imgaug>=0.4.0

# Utilities
tqdm>=4.64.0
click>=8.1.0
rich>=12.0.0
loguru>=0.6.0

# Testing
pytest>=7.1.0
pytest-cov>=3.0.0
pytest-mock>=3.8.0

# Development tools
black>=22.0.0
flake8>=5.0.0
isort>=5.10.0
pre-commit>=2.20.0

# Jupyter notebook support
jupyter>=1.0.0
ipywidgets>=8.0.0
ipykernel>=6.15.0

# Additional scientific libraries
pandas>=1.4.0
xarray>=2022.6.0
dask>=2022.8.0
